--- .eslintignore
@@ -1,8 +1,0 @@
-dist/
-node_modules/
-prisma/migrations/
-*.js
-*.d.ts
-coverage/
-.env
-.env.*
