--- AUTHENTICATION.md
@@ -1,1 +1,0 @@
-
