+++ tests/app.test.ts
@@ -25,24 +25,13 @@ describe('App', () => {
     });
   });
 
-  describe('App structure validation', () => {
-    it('should have proper module structure', () => {
-      // Test basic functionality without creating the full app
-      expect(true).toBe(true);
-    });
-
-    it('should validate environment configuration', () => {
-      // Test that environment variables can be accessed
-      const port = process.env.PORT || 3000;
-      expect(port).toBeDefined();
-    });
-
+  describe('App configuration validation', () => {
     it('should validate OpenAPI configuration structure', () => {
       const openApiConfig = {
         info: {
           title: 'Expense Service API',
           version: '1.0.0',
-          description: 'REST API for managing personal expenses',
+          description: 'A RESTful API for managing expenses',
         },
         servers: [{ url: 'http://localhost:3000' }],
         tags: [
@@ -57,100 +46,33 @@ describe('App', () => {
       expect(openApiConfig.info.version).toBe('1.0.0');
       expect(openApiConfig.servers).toContainEqual({ url: 'http://localhost:3000' });
       expect(openApiConfig.tags.length).toBe(4);
-
-      const tagNames = openApiConfig.tags.map(tag => tag.name);
-      expect(tagNames).toContain('health');
-      expect(tagNames).toContain('auth');
-      expect(tagNames).toContain('users');
-      expect(tagNames).toContain('expenses');
-    });
-  });
-
-  describe('Route structure validation', () => {
-    it('should have health route structure', () => {
-      const healthRoutes = ['/health', '/health/ready'];
-
-      healthRoutes.forEach(route => {
-        expect(route).toMatch(/^\/health/);
-      });
     });
 
-    it('should have auth route structure', () => {
-      const authRoutes = ['/api/v1/auth/register', '/api/v1/auth/login'];
-
-      authRoutes.forEach(route => {
-        expect(route).toMatch(/^\/api\/v1\/auth/);
-      });
-    });
-
-    it('should have user route structure', () => {
-      const userRoutes = ['/api/v1/profile'];
-
-      userRoutes.forEach(route => {
-        expect(route).toMatch(/^\/api\/v1/);
-      });
-    });
-
-    it('should have expense route structure', () => {
-      const expenseRoutes = ['/api/v1/expenses', '/api/v1/expenses/:id'];
-
-      expenseRoutes.forEach(route => {
-        expect(route).toMatch(/^\/api\/v1\/expenses/);
-      });
-    });
-  });
-
-  describe('Error handling structure', () => {
-    it('should handle basic error types', () => {
-      const errorTypes = [
-        { code: 400, message: 'Bad Request' },
-        { code: 401, message: 'Unauthorized' },
-        { code: 404, message: 'Not Found' },
-        { code: 500, message: 'Internal Server Error' },
-      ];
-
-      errorTypes.forEach(error => {
-        expect(error.code).toBeGreaterThanOrEqual(400);
-        expect(error.code).toBeLessThan(600);
-        expect(error.message).toBeDefined();
-        expect(typeof error.message).toBe('string');
-      });
-    });
-
-    it('should validate JSON parsing behavior', () => {
-      const validJson = '{"key": "value"}';
-      const invalidJson = '{"invalid": json}';
-
-      expect(() => JSON.parse(validJson)).not.toThrow();
-      expect(() => JSON.parse(invalidJson)).toThrow();
-    });
-  });
-
-  describe('Configuration validation', () => {
-    it('should have proper Fastify configuration structure', () => {
+    it('should have proper Fastify configuration', () => {
       const fastifyConfig = {
         logger: {
-          level: process.env.LOG_LEVEL || 'info',
+          level: 'info',
+          transport: {
+            target: 'pino-pretty',
+          },
         },
       };
 
       expect(fastifyConfig.logger).toBeDefined();
-      expect(fastifyConfig.logger.level).toBeDefined();
+      expect(fastifyConfig.logger.level).toBe('info');
+      expect(fastifyConfig.logger.transport.target).toBe('pino-pretty');
     });
+  });
 
-    it('should validate plugin registration order', () => {
-      const pluginOrder = [
-        'swagger',
-        'swagger-ui',
-        'sensible',
-        'health-routes',
-        'user-routes',
-        'expense-routes',
-      ];
+  describe('Error handling structure', () => {
+    it('should handle graceful shutdown', () => {
+      // Test graceful shutdown structure
+      const gracefulShutdown = async () => {
+        await mockAppPrisma.$disconnect();
+        return 'shutdown';
+      };
 
-      // Test that we have a logical plugin registration order
-      expect(pluginOrder.indexOf('swagger')).toBeLessThan(pluginOrder.indexOf('swagger-ui'));
-      expect(pluginOrder.indexOf('sensible')).toBeLessThan(pluginOrder.indexOf('health-routes'));
+      expect(typeof gracefulShutdown).toBe('function');
     });
   });
 });
