--- tests/server.test.ts
@@ -1,48 +1,0 @@
-import { describe, it, expect } from '@jest/globals';
-
-describe('Server Entry Point', () => {
-  it('should be defined as a module', () => {
-    // Test that the server module exists and can be imported
-    // We don't want to actually start the server in tests
-    const serverModule = '../src/server';
-    expect(() => require.resolve(serverModule)).not.toThrow();
-  });
-
-  it('should handle environment variables properly', () => {
-    // Test environment variable handling without actually starting server
-    const originalPort = process.env.PORT;
-    const originalHost = process.env.HOST;
-
-    try {
-      process.env.PORT = '4000';
-      process.env.HOST = '127.0.0.1';
-
-      // Verify environment variables are accessible
-      expect(process.env.PORT).toBe('4000');
-      expect(process.env.HOST).toBe('127.0.0.1');
-    } finally {
-      // Restore original values
-      if (originalPort !== undefined) {
-        process.env.PORT = originalPort;
-      } else {
-        delete process.env.PORT;
-      }
-      if (originalHost !== undefined) {
-        process.env.HOST = originalHost;
-      } else {
-        delete process.env.HOST;
-      }
-    }
-  });
-
-  it('should have basic server configuration', () => {
-    // Test basic server configuration constants
-    const defaultPort = 3000;
-    const defaultHost = '0.0.0.0';
-
-    expect(typeof defaultPort).toBe('number');
-    expect(typeof defaultHost).toBe('string');
-    expect(defaultPort).toBeGreaterThan(0);
-    expect(defaultHost.length).toBeGreaterThan(0);
-  });
-});
